fixed the site-settings
